Tested this last night with Hulu and it worked.
I had to run it with pm2 though. Currently if the connection to the AppleTV times out (which happens a bunch), the program will exit. It should know how to recover from this automatically. Not sure if we should recommend running the program with pm2 or build in some fault tolerance.